+++ huginn-proxy-lib/src/proxy/handler/request.rs
@@ -306,7 +306,7 @@ pub async fn handle_proxy_request(
                 hyper::header::HeaderValue::from_str(&fingerprints.ja4_original.raw.to_string())
             {
                 req.headers_mut()
-                    .insert(HeaderName::from_static(names::TLS_JA4_OR), hv);
+                    .insert(HeaderName::from_static(names::TLS_JA4_RO), hv);
             }
             if let Ok(hv) =
                 hyper::header::HeaderValue::from_str(&fingerprints.ja4_stable_v1.full.to_string())
@@ -318,7 +318,7 @@ pub async fn handle_proxy_request(
                 hyper::header::HeaderValue::from_str(&fingerprints.ja4_stable_v1.raw.to_string())
             {
                 req.headers_mut()
-                    .insert(HeaderName::from_static(names::TLS_JA4_S1R), hv);
+                    .insert(HeaderName::from_static(names::TLS_JA4_RS1), hv);
             }
         }
         if let Some(ref rx) = fingerprint_rx {